+++ src/types.ts
@@ -62,6 +62,14 @@ export interface RegistryMap {
  * @property {Object} meta - meta information about the owner of the domain
  */
 export type ResolutionResponse = {
+  ipfs?: {
+    html?: string,
+    redirect_domain ?: string
+  },
+  whois?: {
+    email?: string,
+    for_sale?: boolean
+  },
   addresses: {
     [key: string]: string;
   };
@@ -80,7 +88,7 @@ export type NetworkIdMap = {
   [key: number]: string;
 };
 
-export const UDApiDefaultUrl = "https://unstoppabledomains.com/api/v1";
+export const UDApiDefaultUrl = 'https://unstoppabledomains.com/api/v1';
 
 /**
  * Main configurational object for Resolution instance
@@ -94,11 +102,11 @@ export type Blockchain =
     };
 
 export type API = {
-  url: string
-}
+  url: string;
+};
 
-export const DefaultAPI:API = {
-  url: UDApiDefaultUrl
+export const DefaultAPI: API = {
+  url: UDApiDefaultUrl,
 };
 
 /**
@@ -116,10 +124,10 @@ export type owner = string;
 export type ttl = string;
 export type nodeHash = string;
 export enum NullAddress {
- '0x',
- '0x0000000000000000000000000000000000000000',
- '0x0000000000000000000000000000000000000000000000000000000000000000'
-};
+  '0x',
+  '0x0000000000000000000000000000000000000000',
+  '0x0000000000000000000000000000000000000000000000000000000000000000',
+}
 
 export function isNullAddress(key: string): boolean {
   return Object.values(NullAddress).includes(key);